refactor(ui): split DiscoveryModal device row into its own module
Extract the table row renderer and inline badge/formatter helpers
from DiscoveryModal.tsx (1057 -> 538 lines, clearing the 800-line
red-flag budget). New DiscoveryModalDeviceRow.tsx (514 lines) hosts:

- _methodBadge discovery-method pill
- formatUptime / formatLastSeen / getExpandIcon / getSeverityClasses
- _deviceRow with the expandable LLDP/CDP/SNMP detail panel

The lowercase <methodBadge>, <sortableHeader>, <deviceRow> JSX usages
are pre-existing; this commit only relocates the source, no behavior
change.
