+++ ui/src/components/cards/DiscoveryModal.tsx
@@ -16,32 +16,9 @@
 import type React from 'react';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  button,
-  cn,
-  discoveryMethod as discoveryMethodTheme,
-  icon as iconTokens,
-  modal,
-  radius,
-  severity as severityTheme,
-} from '../../styles/theme';
-import {
-  AlertTriangle,
-  ArrowUpDown,
-  ChevronDown,
-  ChevronUp,
-  Download,
-  RefreshCw,
-  Search,
-  X,
-} from '../ui/icons';
-import { Tooltip } from '../ui/tooltip';
-import type {
-  DiscoveredDevice,
-  DiscoveryMethod,
-  NetworkDiscoveryData,
-  OpenPort,
-} from './NetworkDiscoveryCard';
+import { button, cn, icon as iconTokens, modal, radius } from '../../styles/theme';
+import { ArrowUpDown, ChevronDown, ChevronUp, Download, RefreshCw, Search, X } from '../ui/icons';
+import type { DiscoveredDevice, NetworkDiscoveryData } from './NetworkDiscoveryCard';
 
 interface DiscoveryModalProps {
   isOpen: boolean;
@@ -53,59 +30,6 @@ interface DiscoveryModalProps {
 
 type SortField = 'ip' | 'hostname' | 'vendor' | 'mac' | 'lastSeen';
 type SortDirection = 'asc' | 'desc';
-
-// Discovery method badge
-function _methodBadge({ method }: { method: DiscoveryMethod }): JSX.Element {
-  const theme = discoveryMethodTheme[method] || discoveryMethodTheme.arp;
-  return (
-    <span
-      class={cn('px-1.5 py-0.5 text-xs font-medium uppercase', radius.md, theme.bg, theme.text)}
-    >
-      {method}
-    </span>
-  );
-}
-
-// Format SNMP sysUpTime (in hundredths of a second) to human-readable duration
-function formatUptime(ticks: number): string {
-  const seconds = Math.floor(ticks / 100);
-  const days = Math.floor(seconds / 86400);
-  const hours = Math.floor((seconds % 86400) / 3600);
-  const minutes = Math.floor((seconds % 3600) / 60);
-
-  if (days > 0) {
-    return `${days}d ${hours}h ${minutes}m`;
-  }
-  if (hours > 0) {
-    return `${hours}h ${minutes}m`;
-  }
-  return `${minutes}m`;
-}
-
-// Format timestamp for display
-function formatLastSeen(timestamp: string): string {
-  const date = new Date(timestamp);
-  const now = new Date();
-  const diffMs = now.getTime() - date.getTime();
-  const diffSecs = Math.floor(diffMs / 1000);
-  const diffMins = Math.floor(diffSecs / 60);
-  const diffHours = Math.floor(diffMins / 60);
-  const diffDays = Math.floor(diffHours / 24);
-
-  if (diffSecs < 60) {
-    return 'Just now';
-  }
-  if (diffMins < 60) {
-    return `${diffMins}m ago`;
-  }
-  if (diffHours < 24) {
-    return `${diffHours}h ago`;
-  }
-  if (diffDays < 7) {
-    return `${diffDays}d ago`;
-  }
-  return date.toLocaleDateString();
-}
 
 // Sort comparator
 // biome-ignore lint/complexity/noExcessiveCognitiveComplexity: Comparison function handles multiple sort fields
@@ -149,31 +73,6 @@ function compareDevices(
   return direction === 'asc' ? cmp : -cmp;
 }
 
-// Helper function to get expand icon (avoids nested ternary)
-function getExpandIcon(hasDetails: boolean, isExpanded: boolean): string {
-  if (!hasDetails) {
-    return '';
-  }
-  if (isExpanded) {
-    return '▲';
-  }
-  return '▼';
-}
-
-// Helper function to get severity theme classes (avoids nested ternary)
-function getSeverityClasses(severity: string): string {
-  if (severity === 'CRITICAL') {
-    return `${severityTheme.critical.bg} ${severityTheme.critical.text}`;
-  }
-  if (severity === 'HIGH') {
-    return `${severityTheme.high.bg} ${severityTheme.high.text}`;
-  }
-  if (severity === 'MEDIUM') {
-    return `${severityTheme.medium.bg} ${severityTheme.medium.text}`;
-  }
-  return `${severityTheme.low.bg} ${severityTheme.low.text}`;
-}
-
 // Helper function to get sort icon (avoids nested ternary)
 function getSortIcon(isActive: boolean, direction: SortDirection): JSX.Element {
   if (!isActive) {
@@ -215,424 +114,6 @@ function _sortableHeader({
         {getSortIcon(isActive, direction)}
       </div>
     </th>
-  );
-}
-
-// Device row component
-// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: Device row handles many device types and states
-function _deviceRow({
-  device,
-  isExpanded,
-  onToggle,
-  onDeepScan,
-  isScanning,
-}: {
-  device: DiscoveredDevice;
-  isExpanded: boolean;
-  onToggle: () => void;
-  onDeepScan?: (ip: string) => Promise<void>;
-  isScanning: boolean;
-}): JSX.Element {
-  const { t } = useTranslation('cards');
-  const openPorts = device.profile?.openPorts?.filter((p) => p.isOpen) || [];
-  const hasDetails =
-    device.lldpInfo ||
-    device.cdpInfo ||
-    device.edpInfo ||
-    device.ndpInfo ||
-    device.snmpData ||
-    openPorts.length > 0;
-
-  const handleScan = async (e: React.MouseEvent): Promise<void> => {
-    e.stopPropagation();
-    if (onDeepScan && device.ip) {
-      // biome-ignore lint/nursery/useAwaitThenable: onDeepScan returns Promise<void>
-      await onDeepScan(device.ip);
-    }
-  };
-
-  return (
-    <>
-      <tr
-        class={cn(
-          'border-b border-surface-border hover:bg-surface-hover cursor-pointer transition-colors',
-          isExpanded && 'bg-surface-hover',
-        )}
-        onClick={onToggle}
-      >
-        {/* IP Address */}
-        <td class="px-3 py-2">
-          <div class="flex flex-col">
-            <span class="font-mono text-sm font-medium text-text-primary">
-              {device.ip || t('network.noIP')}
-            </span>
-            {device.ipv6 ? (
-              <span class="font-mono text-xs text-text-muted truncate max-w-40" title={device.ipv6}>
-                {device.ipv6.length > 25 ? `${device.ipv6.substring(0, 25)}...` : device.ipv6}
-              </span>
-            ) : null}
-          </div>
-        </td>
-
-        {/* Hostname - prefer displayName, fallback to mdnsName, netbiosName, hostname */}
-        <td class="px-3 py-2">
-          <span
-            class="text-sm text-text-secondary truncate block max-w-40"
-            title={device.displayName || device.mdnsName || device.netbiosName || device.hostname}
-          >
-            {device.displayName || device.mdnsName || device.netbiosName || device.hostname || '-'}
-          </span>
-        </td>
-
-        {/* MAC Address */}
-        <td class="px-3 py-2">
-          <span class="font-mono text-xs text-text-muted">{device.mac || '-'}</span>
-        </td>
-
-        {/* Vendor */}
-        <td class="px-3 py-2">
-          {device.vendor === 'LAA' ? (
-            <Tooltip
-              content="Locally Administered Address - MAC assigned locally rather than by manufacturer"
-              position="bottom"
-            >
-              <span class="text-xs text-text-muted underline decoration-dotted cursor-help">
-                LAA
-              </span>
-            </Tooltip>
-          ) : (
-            <span class="text-xs text-text-muted truncate block max-w-28" title={device.vendor}>
-              {device.vendor || '-'}
-            </span>
-          )}
-        </td>
-
-        {/* Discovery Methods */}
-        <td class="px-3 py-2">
-          <div class="flex items-center gap-1 flex-wrap">
-            {device.discoveryMethod.map((method) => (
-              <methodBadge key={method} method={method} />
-            ))}
-          </div>
-        </td>
-
-        {/* Open Ports */}
-        <td class="px-3 py-2">
-          {openPorts.length > 0 ? (
-            <span
-              class={cn(
-                'text-xs px-1.5 py-0.5 bg-status-success/20 text-status-success',
-                radius.md,
-              )}
-            >
-              {openPorts.length} open
-            </span>
-          ) : (
-            <span class="text-xs text-text-muted">-</span>
-          )}
-        </td>
-
-        {/* Vulnerabilities */}
-        <td class="px-3 py-2">
-          {device.vulnerabilities && device.vulnerabilities.count > 0 ? (
-            <span
-              class={cn(
-                'inline-flex items-center gap-1 text-xs px-1.5 py-0.5',
-                radius.md,
-                getSeverityClasses(device.vulnerabilities.highestSeverity),
-              )}
-            >
-              <AlertTriangle class="w-3 h-3" />
-              {device.vulnerabilities.count}
-            </span>
-          ) : (
-            <span class="text-xs text-text-muted">-</span>
-          )}
-        </td>
-
-        {/* Last Seen */}
-        <td class="px-3 py-2">
-          <span class="text-xs text-text-muted">{formatLastSeen(device.lastSeen)}</span>
-        </td>
-
-        {/* Actions */}
-        <td class="px-3 py-2">
-          <div class="flex items-center gap-2">
-            {onDeepScan && device.ip ? (
-              <button
-                type="button"
-                onClick={handleScan}
-                disabled={isScanning}
-                class={cn(
-                  'text-xs px-2 py-1 bg-brand-primary/20 text-brand-primary',
-                  radius.md,
-                  'hover:bg-brand-primary/30 transition-colors disabled:opacity-50',
-                )}
-              >
-                {isScanning ? '...' : t('discovery.scan')}
-              </button>
-            ) : null}
-            <span class="text-xs text-text-muted">{getExpandIcon(hasDetails, isExpanded)}</span>
-          </div>
-        </td>
-      </tr>
-
-      {/* Expanded details row */}
-      {isExpanded && hasDetails ? (
-        <tr class="bg-surface-sunken">
-          <td colSpan={9} class="px-4 py-3">
-            <div class="space-y-3">
-              {/* Open Ports */}
-              {openPorts.length > 0 ? (
-                <div>
-                  <h4 class="text-xs font-semibold text-text-secondary mb-1">Open Ports</h4>
-                  <div class="flex flex-wrap gap-2">
-                    {openPorts.map((port: OpenPort) => (
-                      <span
-                        key={port.port}
-                        class={cn(
-                          'px-2 py-1 text-xs font-mono',
-                          radius.md,
-                          'bg-surface-base text-text-primary',
-                        )}
-                      >
-                        {port.port}/{port.protocol}{' '}
-                        {port.service ? (
-                          <span class="text-text-muted">({port.service})</span>
-                        ) : null}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              ) : null}
-
-              {/* LLDP Info */}
-              {device.lldpInfo ? (
-                <div>
-                  <h4 class="text-xs font-semibold text-text-secondary mb-1">LLDP Information</h4>
-                  <div class="grid grid-cols-2 md:grid-cols-4 gap-2 text-xs">
-                    <div>
-                      <span class="text-text-muted">System:</span> {device.lldpInfo.systemName}
-                    </div>
-                    <div>
-                      <span class="text-text-muted">Port:</span> {device.lldpInfo.portId}
-                    </div>
-                    {device.lldpInfo.managementAddress ? (
-                      <div>
-                        <span class="text-text-muted">Mgmt IP:</span>{' '}
-                        {device.lldpInfo.managementAddress}
-                      </div>
-                    ) : null}
-                    {device.lldpInfo.capabilities ? (
-                      <div>
-                        <span class="text-text-muted">Capabilities:</span>{' '}
-                        {device.lldpInfo.capabilities.join(', ')}
-                      </div>
-                    ) : null}
-                  </div>
-                </div>
-              ) : null}
-
-              {/* CDP Info */}
-              {device.cdpInfo ? (
-                <div>
-                  <h4 class="text-xs font-semibold text-text-secondary mb-1">CDP Information</h4>
-                  <div class="grid grid-cols-2 md:grid-cols-4 gap-2 text-xs">
-                    <div>
-                      <span class="text-text-muted">Device:</span> {device.cdpInfo.deviceId}
-                    </div>
-                    <div>
-                      <span class="text-text-muted">Platform:</span> {device.cdpInfo.platform}
-                    </div>
-                    {device.cdpInfo.nativeVlan ? (
-                      <div>
-                        <span class="text-text-muted">Native VLAN:</span>{' '}
-                        {device.cdpInfo.nativeVlan}
-                      </div>
-                    ) : null}
-                  </div>
-                </div>
-              ) : null}
-
-              {/* SNMP Data */}
-              {device.snmpData ? (
-                <div class="space-y-2">
-                  <h4 class="text-xs font-semibold text-text-secondary">
-                    {t('discovery.snmpInfo', 'SNMP Details')}
-                  </h4>
-
-                  {/* System Info */}
-                  {device.snmpData.system ? (
-                    <div class="bg-surface-base p-2 rounded-md">
-                      <div class="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-x-4 gap-y-1 text-xs">
-                        {device.snmpData.system.sysName ? (
-                          <div>
-                            <span class="text-text-muted">Name:</span>{' '}
-                            <span class="text-text-primary font-medium">
-                              {device.snmpData.system.sysName}
-                            </span>
-                          </div>
-                        ) : null}
-                        {device.snmpData.system.sysDescr ? (
-                          <div class="col-span-2">
-                            <span class="text-text-muted">Description:</span>{' '}
-                            <span class="text-text-primary">
-                              {device.snmpData.system.sysDescr.length > 80
-                                ? `${device.snmpData.system.sysDescr.substring(0, 80)}...`
-                                : device.snmpData.system.sysDescr}
-                            </span>
-                          </div>
-                        ) : null}
-                        {device.snmpData.system.sysLocation ? (
-                          <div>
-                            <span class="text-text-muted">Location:</span>{' '}
-                            <span class="text-text-primary">
-                              {device.snmpData.system.sysLocation}
-                            </span>
-                          </div>
-                        ) : null}
-                        {device.snmpData.system.sysContact ? (
-                          <div>
-                            <span class="text-text-muted">Contact:</span>{' '}
-                            <span class="text-text-primary">
-                              {device.snmpData.system.sysContact}
-                            </span>
-                          </div>
-                        ) : null}
-                        {device.snmpData.system.sysUpTime !== undefined &&
-                        device.snmpData.system.sysUpTime > 0 ? (
-                          <div>
-                            <span class="text-text-muted">Uptime:</span>{' '}
-                            <span class="text-text-primary">
-                              {formatUptime(device.snmpData.system.sysUpTime)}
-                            </span>
-                          </div>
-                        ) : null}
-                      </div>
-                    </div>
-                  ) : null}
-
-                  {/* Interfaces Summary */}
-                  {device.snmpData.interfaces && device.snmpData.interfaces.length > 0 ? (
-                    <div>
-                      <span class="text-xs text-text-muted">
-                        Interfaces ({device.snmpData.interfaces.length}):
-                      </span>
-                      <div class="flex flex-wrap gap-1 mt-1">
-                        {/* biome-ignore lint/complexity/noExcessiveCognitiveComplexity: Complex interface status rendering */}
-                        {device.snmpData.interfaces.slice(0, 8).map((iface) => (
-                          <span
-                            key={iface.name}
-                            class={cn(
-                              'px-1.5 py-0.5 text-xs',
-                              radius.sm,
-                              iface.operStatus === 'up'
-                                ? 'bg-status-success/20 text-status-success'
-                                : 'bg-surface-hover text-text-muted',
-                            )}
-                            title={`${iface.name} - ${iface.speed ? `${Math.round(iface.speed / 1000000)} Mbps` : 'N/A'}`}
-                          >
-                            {iface.name}
-                            {iface.speed && iface.speed > 0 ? (
-                              <span class="text-text-muted ml-1">
-                                {iface.speed >= 1000000000
-                                  ? `${Math.round(iface.speed / 1000000000)}G`
-                                  : `${Math.round(iface.speed / 1000000)}M`}
-                              </span>
-                            ) : null}
-                          </span>
-                        ))}
-                        {device.snmpData.interfaces.length > 8 ? (
-                          <span class="text-xs text-text-muted">
-                            +{device.snmpData.interfaces.length - 8} more
-                          </span>
-                        ) : null}
-                      </div>
-                    </div>
-                  ) : null}
-
-                  {/* VLANs Summary */}
-                  {device.snmpData.vlans && device.snmpData.vlans.length > 0 ? (
-                    <div>
-                      <span class="text-xs text-text-muted">
-                        VLANs ({device.snmpData.vlans.length}):
-                      </span>
-                      <div class="flex flex-wrap gap-1 mt-1">
-                        {device.snmpData.vlans.slice(0, 12).map((vlan) => (
-                          <span
-                            key={vlan.id}
-                            class={cn(
-                              'px-1.5 py-0.5 text-xs bg-brand-primary/10 text-brand-primary',
-                              radius.sm,
-                            )}
-                            title={vlan.name || `VLAN ${vlan.id}`}
-                          >
-                            {vlan.id}
-                            {vlan.name && vlan.name !== `VLAN${vlan.id}` ? (
-                              <span class="text-text-muted ml-1">
-                                {vlan.name.length > 10
-                                  ? `${vlan.name.substring(0, 10)}...`
-                                  : vlan.name}
-                              </span>
-                            ) : null}
-                          </span>
-                        ))}
-                        {device.snmpData.vlans.length > 12 ? (
-                          <span class="text-xs text-text-muted">
-                            +{device.snmpData.vlans.length - 12} more
-                          </span>
-                        ) : null}
-                      </div>
-                    </div>
-                  ) : null}
-
-                  {/* Hardware Inventory */}
-                  {device.snmpData.entities && device.snmpData.entities.length > 0 ? (
-                    <div>
-                      <span class="text-xs text-text-muted">Hardware:</span>
-                      <div class="grid grid-cols-1 md:grid-cols-2 gap-1 mt-1 text-xs">
-                        {device.snmpData.entities
-                          .filter(
-                            (e) =>
-                              e.physicalClass === 'chassis' ||
-                              e.physicalClass === 'module' ||
-                              e.physicalClass === 'powerSupply',
-                          )
-                          .slice(0, 4)
-                          .map((entity) => (
-                            <div
-                              key={entity.serialNum || entity.name || entity.description}
-                              class="bg-surface-hover px-2 py-1 rounded"
-                            >
-                              <span class="text-text-primary">
-                                {entity.name || entity.description}
-                              </span>
-                              {entity.serialNum ? (
-                                <span class="text-text-muted ml-2">S/N: {entity.serialNum}</span>
-                              ) : null}
-                              {entity.modelName ? (
-                                <span class="text-text-muted ml-2">Model: {entity.modelName}</span>
-                              ) : null}
-                            </div>
-                          ))}
-                      </div>
-                    </div>
-                  ) : null}
-                </div>
-              ) : null}
-
-              {/* OS Guess */}
-              {device.osGuess ? (
-                <div>
-                  <span class="text-xs text-text-muted">OS Guess:</span>{' '}
-                  <span class="text-xs text-text-primary">{device.osGuess}</span>
-                </div>
-              ) : null}
-            </div>
-          </td>
-        </tr>
-      ) : null}
-    </>
   );
 }
